requirements.txt
Flask==2.3.3
Werkzeug==2.3.8
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
SQLAlchemy==2.0.20
python-dotenv==1.0.0
gunicorn==21.2.0
email-validator==2.0.0.post2
WTForms==3.1.1
dnspython==2.6.1

